feat: refactored token table
