+++ app/components/faucet.js
@@ -1,17 +1,13 @@
-import Image from "next/image";
 import { useState } from "react";
 import ReCAPTCHA from "react-google-recaptcha";
 import {
   ConnectWallet,
   lightTheme,
-  useChainId,
-  useAddress,
   useConnectionStatus,
   useWallet
 } from "@thirdweb-dev/react";
 
-import { ClaimButton } from "./claimButton";
-import { AddTokenToWalletButton } from "./addTokenToWallet"
+import { TokenButtonsTable } from "./TokenButtonsTable";
 
 const customTheme = lightTheme({
   colors: {
@@ -27,9 +23,7 @@ const customTheme = lightTheme({
 const Faucet = () => {
   const [captchaCompleted, setCaptchaCompleted] = useState(false);
 
-  const address = useAddress();
   const walletStatus = useConnectionStatus();
-  const chainId = useChainId();
   const wallet = useWallet();
 
   const tokens = [
@@ -53,87 +47,6 @@ const Faucet = () => {
     )
   }
 
-  const TokenButtonsTable = ({ tokens, walletStatus, captchaCompleted, wallet }) => {
-    return (
-      <div className="overflow-x-auto">
-        <table className="w-full divide-y divide-zinc-600 bg-black-800 text-center">
-          <thead className="bg-black-800">
-            <tr>
-              <th
-                scope="col"
-                className="px-6 text-center text-xs font-medium text-white uppercase tracking-wider"
-              >
-                Image
-              </th>
-              <th
-                scope="col"
-                className="px-6 text-center text-xs font-medium text-white uppercase tracking-wider"
-              >
-                Name
-              </th>
-              <th
-                scope="col"
-                className="px-6 text-center text-xs font-medium text-white uppercase tracking-wider"
-              >
-                Symbol
-              </th>
-              <th
-                scope="col"
-                className="px-6 text-center text-xs font-medium text-white uppercase tracking-wider"
-              >
-                Address
-              </th>
-              <th
-                scope="col"
-                className="px-6 text-center text-xs font-medium text-white uppercase tracking-wider"
-              >
-                Action
-              </th>
-              {wallet.walletId === 'metamask' &&
-                <th
-                  scope="col"
-                  className="px-6 text-center text-xs font-medium text-white uppercase tracking-wider"
-                >
-                  Add To Wallet
-                </th>
-              }
-            </tr>
-          </thead>
-          <tbody className="bg-black-800 divide-y divide-zinc-600">
-            {tokens.map((token, index) => (
-              <tr key={index} className="bg-black-800">
-                <td className="px-6 text-sm font-medium text-white">
-                  <Image
-                    src={token.logo}
-                    alt=''
-                    width={32}
-                    height={32}
-                  />
-                </td>
-                <td className="px-6  text-sm font-medium text-white">{token.name}</td>
-                <td className="px-6 whitespace-nowrap text-sm text-white">{token.symbol}</td>
-                <td className="px-6 whitespace-nowrap text-sm text-white">{token.address}</td>
-                <td className="px-6 whitespace-nowrap text-sm text-white flex justify-center items-center mt-1">
-                  <ClaimButton
-                    tokenAddress={token.address}
-                    walletStatus={walletStatus}
-                    captchaCompleted={captchaCompleted}
-                    chainId={chainId}
-                    address={address}
-                    amount={token.amount}
-                  />
-                </td>
-                {wallet.walletId === 'metamask' && token.name !== 'Tezos' &&
-                <td className="px-6 whitespace-nowrap text-sm text-white">
-                  <AddTokenToWalletButton token={token} />
-                </td>}
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    )
-  };
 
   return (
     <div className="flex items-center justify-center w-full lg:w-1/2 min-w-1/2 rounded-lg mt-10 mb-10">
